+++ NEWS.md
@@ -1,5 +1,6 @@
 # rigakuXRD 0.2.3
 
+* update RAS file reader
 * reading of ASC, TXT, RAS, and RASX files should yield the same columns, when using `xrd.import()`, unify columns
 
 # rigakuXRD 0.2.2